requirements.txt
PyPDF2
tenacity
pytesseract
pdfplumber
pdf2image
pandas
pillow
opencv-python
requests
matplotlib
pymupdf
numpy
openai
google-cloud-vision
unstructured
pdfminer.six
pillow-heif
matplotlib
pi-heif
unstructured-inference
unstructured-pytesseract